+++ styles/index.ts
@@ -2,3 +2,4 @@
 import '@slidev/client/styles/layouts-base.css'
 import './layout.css'
 import './code.css'
+import './mermaid.css'