+++ changelog.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [0.0.4] - 2023-02-02
+
+- Added mermaid styles for `sequenceDiagram`
+- Small fixes
+
+## [0.0.3] - 2023-02-02
+
+- Updated styles for all layouts
+- Added screenshots and documentation
+
 ## [0.0.2] - 2023-01-31
 
 - New layouts